♻️ refactor(shared): reorganize shared UI and library exports
