+++ src/shared/ui/icons/animated-check.stories.tsx
