+++ src/shared/ui/icons/animated-check.tsx
