+++ src/shared/ui/icons/asterisk.stories.tsx
