+++ src/shared/ui/icons/asterisk.tsx
