+++ src/shared/ui/icons/figma.stories.tsx
