+++ src/shared/ui/icons/figma.tsx
