+++ src/shared/ui/icons/github.stories.tsx
