+++ src/shared/ui/icons/icons.test.tsx
