+++ src/shared/ui/icons/index.ts
