+++ src/shared/ui/icons/koutyuke.stories.tsx
