+++ src/shared/ui/icons/koutyuke.tsx
