+++ src/shared/ui/icons/lucide-props.ts
