+++ src/shared/ui/icons/type.ts
