+++ src/shared/ui/icons/zenn.stories.tsx
