+++ src/shared/ui/section-divider/index.ts
@@ -1,0 +1,1 @@
+export { SectionDivider } from "./section-divider";
