+++ src/shared/ui/section-divider/section-divider.tsx
@@ -1,4 +1,4 @@
-import { cn } from "../../lib/cn";
+import { cn } from "../../lib";
 
 type SectionDividerProps = {
   className?: string;
