+++ src/shared/ui/section-heading/index.ts
@@ -1,0 +1,1 @@
+export { SectionHeading } from "./section-heading";
