+++ src/shared/ui/section-heading/section-heading.stories.tsx
