+++ src/shared/ui/section-heading/section-heading.tsx
@@ -1,4 +1,4 @@
-import { cn } from "../../lib/cn";
+import { cn } from "../../lib";
 
 type SectionHeadingProps = {
   align?: "center" | "start";
