+++ src/shared/ui/tag-list/index.ts
@@ -1,0 +1,1 @@
+export { TagList } from "./tag-list";
