+++ src/shared/ui/tag-list/tag-list.tsx
@@ -1,4 +1,4 @@
-import { cn } from "../../lib/cn";
+import { cn } from "../../lib";
 
 type TagListProps = {
   className?: string;
